moved parse_config.rs to ser_de/config
