+++ src/ser_de/manifest/mod.rs
@@ -1,0 +1,1 @@
+pub mod manifest_utils;
