+++ src/ser_de/mod.rs
@@ -1,0 +1,2 @@
+pub mod config;
+mod manifest;
