+++ src/utils/mod.rs
@@ -1,4 +1,3 @@
 pub mod do_vectors_match;
-pub mod manifest_utils;
-pub mod config;
 pub mod fs_recursive_copy;
+pub mod parse_config;
